backend: refactoring. create google_cds_store package
